+++ src/main/kotlin/net/liquidev/dawd3/ui/widget/Widget.kt
@@ -4,28 +4,28 @@ import net.liquidev.dawd3.ui.Event
 import net.liquidev.dawd3.ui.EventContext
 import net.minecraft.client.util.math.MatrixStack
 
-abstract class Widget(var x: Int, var y: Int) {
-    abstract val width: Int
-    abstract val height: Int
+abstract class Widget(var x: Float, var y: Float) {
+    abstract val width: Float
+    abstract val height: Float
 
     protected abstract fun drawContent(
         matrices: MatrixStack,
-        mouseX: Int,
-        mouseY: Int,
+        mouseX: Float,
+        mouseY: Float,
         deltaTime: Float,
     )
 
     /** Returns false to propagate the event, or true to consume it. */
     abstract fun event(context: EventContext, event: Event): Boolean
 
-    fun draw(matrices: MatrixStack, mouseX: Int, mouseY: Int, deltaTime: Float) {
+    fun draw(matrices: MatrixStack, mouseX: Float, mouseY: Float, deltaTime: Float) {
         matrices.push()
         matrices.translate(x.toDouble(), y.toDouble(), 0.0)
         drawContent(matrices, mouseX - x, mouseY - y, deltaTime)
         matrices.pop()
     }
 
-    fun containsRelativePoint(x: Int, y: Int) =
+    fun containsRelativePoint(x: Float, y: Float) =
         x >= 0 && y >= 0 && x <= width && y <= height
 
     companion object {
@@ -39,7 +39,7 @@ abstract class Widget(var x: Int, var y: Int) {
                 if (
                     widget.event(
                         context,
-                        event.relativeTo(widget.x.toDouble(), widget.y.toDouble())
+                        event.relativeTo(widget.x, widget.y)
                     )
                 ) {
                     return true
